calling fetch inside download button handler
